Distance grab randomly works on the bottle.
now the left hand is grabber while the right hand is distance grabber
The ball is grabbable, while the bottle is distance grabbable

Both hands can grab the bottle. The right hand randomly grabs the bottle at a distance. Unclear what exactly triggers it but it suddenly snaps to my hands. There is also a light glow around the bottle if I move my right hand, which does not happen if I move my left hand.

Left hand can grab the ball, but right hand seems to create some explosive force on the ball.

(also, made the car bigger, and moved the items to the edge of the car, and disable the obstacles + movement of the car for ease of use)